Move puppeteer helpers in __tests__
It separates our test helpers more clearly from helpers for the actual site
